Here is the concise commit message summarizing the changes:
Optimize entitlement checks, message processing, and update dependencies

- Refactored entitlement logic to support bulk fetching for text and image models, significantly reducing database queries.
- Updated `RequestLedger.cost` to `float` and adjusted usage calculation to sum costs directly.
- Optimized `generate_and_save_title` to skip generation if the conversation title is already set.
- Implemented content caching in `_upsert_text` and eager loading for message history to improve performance during streaming and retrieval.
- Refactored chat entitlement verification and error handling in `chat_helpers.py`.
- Added optimization to skip S3 metadata checks for derived images in `image_deriver.py`.
- Updated project dependencies including `openai`, `sqlmodel`, and `pillow`.
- bumped up the version number

+++ app/db/models.py
@@ -218,7 +218,7 @@ class RequestLedger(SQLModel, table=True):
     request_id: str = Field(index=True)  # client- or server-generated; used for idempotency
     model_name: str = Field(index=True)
     feature: str = Field(index=True)
-    cost: int = Field(default=1)
+    cost: float = Field(default=1.0)
 
     state: State = Field(default=State.reserved, index=True)
     tool_choice: Optional[str] = None     # e.g., "auto" or "image_generation"
